# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: POS Platform API
  version: 0.1.0
  description: >
    REST API for the POS platform providing checkout, backoffice, and
    synchronization capabilities. The API is designed to be idempotent, secure,
    and ready for hybrid deployments.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/product-groups:
    get:
      tags:
        - Catalog
      summary: List product groups
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Product groups ordered by name.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductGroup'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags:
        - Catalog
      summary: Create product group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductGroupWrite'
      responses:
        '201':
          description: Product group created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductGroup'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/product-groups/{id}:
    get:
      tags:
        - Catalog
      summary: Get product group
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Product group.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductGroup'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    patch:
      tags:
        - Catalog
      summary: Update product group
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductGroupPatch'
      responses:
        '200':
          description: Product group updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductGroup'
        '204':
          description: No fields changed.
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/products:
    get:
      tags:
        - Catalog
      summary: List products
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Products ordered by name.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags:
        - Catalog
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductWrite'
      responses:
        '201':
          description: Product created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/products/{id}:
    get:
      tags:
        - Catalog
      summary: Get product
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Product.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    patch:
      tags:
        - Catalog
      summary: Update product
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductPatch'
      responses:
        '200':
          description: Product updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '204':
          description: No fields changed.
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/customers:
    get:
      tags:
        - Customers
      summary: List customers
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Customers ordered by name.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags:
        - Customers
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerWrite'
      responses:
        '201':
          description: Customer created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/customers/{id}:
    get:
      tags:
        - Customers
      summary: Get customer
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Customer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    patch:
      tags:
        - Customers
      summary: Update customer
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerPatch'
      responses:
        '200':
          description: Customer updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '204':
          description: No fields changed.
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /api/v1/payment-terms:
    get:
      tags:
        - Catalog
      summary: List payment terms
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Payment terms ordered by name.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentTerm'
    post:
      tags:
        - Catalog
      summary: Create payment term
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentTermWrite'
      responses:
        '201':
          description: Payment term created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentTerm'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/payment-terms/{id}:
    get:
      tags:
        - Catalog
      summary: Get payment term
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Payment term.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentTerm'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    patch:
      tags:
        - Catalog
      summary: Update payment term
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentTermPatch'
      responses:
        '200':
          description: Payment term updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentTerm'
        '204':
          description: No fields changed.
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SearchQuery:
      name: search
      in: query
      description: Case insensitive partial match on name, legacy code, or identifier.
      schema:
        type: string
    LimitQuery:
      name: limit
      in: query
      description: Maximum number of records to return (default 100, max 500).
      schema:
        type: integer
        minimum: 1
        maximum: 500
  responses:
    ValidationError:
      description: Request payload failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                message: resource not found
    ConflictError:
      description: Unique constraint violation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                message: conflict
    UnauthorizedError:
      description: Authentication credentials are missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                message: unauthorized
    ForbiddenError:
      description: Authenticated user lacks permission for the requested action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                message: forbidden
  schemas:
    ProductGroup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        legacy_code:
          type: string
          nullable: true
        name:
          type: string
      required:
        - id
        - name
    ProductGroupWrite:
      type: object
      properties:
        legacy_code:
          type: string
          nullable: true
        name:
          type: string
      required:
        - name
    ProductGroupPatch:
      allOf:
        - $ref: '#/components/schemas/ProductGroupWrite'
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        legacy_code:
          type: string
          nullable: true
        name:
          type: string
        barcode:
          type: string
          nullable: true
        group_id:
          type: string
          format: uuid
          nullable: true
        reference:
          type: string
          nullable: true
        min_stock:
          type: number
          format: float
          nullable: true
        price_cash:
          type: number
          format: float
          nullable: true
        price_base:
          type: number
          format: float
          nullable: true
      required:
        - id
        - name
    ProductWrite:
      type: object
      properties:
        legacy_code:
          type: string
          nullable: true
        name:
          type: string
        barcode:
          type: string
          nullable: true
        group_id:
          type: string
          format: uuid
          nullable: true
        reference:
          type: string
          nullable: true
        min_stock:
          type: number
          format: float
          nullable: true
        price_cash:
          type: number
          format: float
          nullable: true
        price_base:
          type: number
          format: float
          nullable: true
      required:
        - name
    ProductPatch:
      allOf:
        - $ref: '#/components/schemas/ProductWrite'
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        legacy_code:
          type: string
          nullable: true
        name:
          type: string
        cpf:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        uf:
          type: string
          nullable: true
        cep:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - active
            - delinquent
            - inactive
        credit_limit:
          type: number
          format: float
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - id
        - name
        - status
    CustomerWrite:
      type: object
      properties:
        legacy_code:
          type: string
          nullable: true
        name:
          type: string
        cpf:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        uf:
          type: string
          nullable: true
          minLength: 2
          maxLength: 2
        cep:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - active
            - delinquent
            - inactive
          default: active
        credit_limit:
          type: number
          format: float
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name
    CustomerPatch:
      allOf:
        - $ref: '#/components/schemas/CustomerWrite'
...
    PaymentTerm:
      type: object
      properties:
        id:
          type: string
          format: uuid
        legacy_code:
          type: string
          nullable: true
        name:
          type: string
      required:
        - id
        - name
    PaymentTermWrite:
      type: object
      properties:
        legacy_code:
          type: string
          nullable: true
        name:
          type: string
      required:
        - name
    PaymentTermPatch:
      allOf:
        - $ref: '#/components/schemas/PaymentTermWrite'
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          nullable: true
      required:
        - message
    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: validation_failed
        issues:
          type: array
          items:
            type: object
            properties:
              path:
                type: array
                items:
                  type: string
              message:
                type: string
              code:
                type: string
            required:
              - path
              - message
              - code
      required:
        - message
        - issues
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT